+++ src/IDListLayer.hpp
@@ -3,7 +3,7 @@
 
 using namespace geode::prelude;
 
-class IDListLayer : public CCLayer, LevelManagerDelegate {
+class IDListLayer : public CCLayer, SetIDPopupDelegate, LevelManagerDelegate, TextInputDelegate {
 public:
     inline static std::vector<int> AREDL = {};
     inline static std::vector<std::string> AREDL_NAMES = {};
@@ -20,6 +20,11 @@ public:
     void onRight(CCObject*);
     void onInfo(CCObject*);
     void onRefresh(CCObject*);
+    void onPage(CCObject*);
+    void onRandom(CCObject*);
+    void onFirst(CCObject*);
+    void onLast(CCObject*);
+    void keyDown(enumKeyCodes) override;
 protected:
     GJListLayer* m_list;
     CCMenu* m_backMenu;
@@ -32,6 +37,9 @@ protected:
     CCLayerColor* m_searchBarView;
     CCTextInputNode* m_searchBar;
     CCLabelBMFont* m_countLabel;
+    CCLabelBMFont* m_pageLabel;
+    CCMenuItemSpriteExtra* m_firstButton;
+    CCMenuItemSpriteExtra* m_lastButton;
     int m_page = 0;
     std::string m_query = "";
     std::vector<int> m_fullSearchResults;
@@ -48,4 +56,14 @@ protected:
         loadLevelsFailed(test);
     }
     void setupPageInfo(std::string, const char*) {}
+    void textChanged(CCTextInputNode*) {}
+    void textInputOpened(CCTextInputNode*) {}
+    void textInputClosed(CCTextInputNode*) {}
+    void textInputShouldOffset(CCTextInputNode*, float) {}
+    void textInputReturn(CCTextInputNode*) {}
+    bool allowTextInput(CCTextInputNode*) {
+        return true;
+    }
+    void enterPressed(CCTextInputNode*);
+    void setIDPopupClosed(SetIDPopup*, int);
 };